PickSecurityLevel refactor, Disclaimer component
