+++ src/components/guard/setup/ServerSetupCard/components/PickSecurityLevel/index.ts
@@ -1,0 +1,3 @@
+import PickSecurityLevel from "./PickSecurityLevel"
+
+export default PickSecurityLevel
